+++ engine/src/vulkan/module/overlay/OverlayUniformHandler.hpp
@@ -91,7 +91,7 @@ cOverlayUniformHandler::cOverlayUniformHandler(cLogicalDevice* pLogicalDevice, c
 
     VkDescriptorSetLayoutCreateInfo tLayoutInfo = {};
     tLayoutInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO;
-    tLayoutInfo.bindingCount = atBindings.size();
+    tLayoutInfo.bindingCount = (uint)atBindings.size();
     tLayoutInfo.pBindings = atBindings.data();
 
     if (!pLogicalDevice->CreateDescriptorSetLayout(&tLayoutInfo, nullptr, &poDescriptorSetLayout))
@@ -118,7 +118,7 @@ cOverlayUniformHandler::cOverlayUniformHandler(cLogicalDevice* pLogicalDevice, c
 
     VkDescriptorSetLayoutCreateInfo tElementLayoutInfo = {};
     tElementLayoutInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO;
-    tElementLayoutInfo.bindingCount = atElementBindings.size();
+    tElementLayoutInfo.bindingCount = (uint)atElementBindings.size();
     tElementLayoutInfo.pBindings = atElementBindings.data();
 
     if (!pLogicalDevice->CreateDescriptorSetLayout(&tElementLayoutInfo, nullptr, &poElementDescriptorSetLayout))
@@ -161,7 +161,7 @@ void cOverlayUniformHandler::CreateUniformBuffers(cScene* pScene)
                                 poBuffer, poBufferMemory);
 
     VkDeviceSize bufferSize = sizeof(tOverlayElementObject);
-    uint uiCount = pScene->GetOverlay().size();
+    uint uiCount = (uint)pScene->GetOverlay().size();
 
     paoElementUniformBuffers.resize(uiCount);
     paoElementUniformBuffersMemory.resize(uiCount);
@@ -178,16 +178,16 @@ void cOverlayUniformHandler::CreateDescriptorPool()
 {
     std::array<VkDescriptorPoolSize, 2> atPoolSizes = {};
     atPoolSizes[0].type = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-    atPoolSizes[0].descriptorCount = 1 + paoElementUniformBuffers.size();
+    atPoolSizes[0].descriptorCount = 1 + (uint)paoElementUniformBuffers.size();
     atPoolSizes[1].type = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-    atPoolSizes[1].descriptorCount = 1 + paoElementUniformBuffers.size();
+    atPoolSizes[1].descriptorCount = 1 + (uint)paoElementUniformBuffers.size();
 
     VkDescriptorPoolCreateInfo tPoolInfo = {};
     tPoolInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
-    tPoolInfo.poolSizeCount = atPoolSizes.size();
+    tPoolInfo.poolSizeCount = (uint)atPoolSizes.size();
     tPoolInfo.pPoolSizes = atPoolSizes.data();
 
-    tPoolInfo.maxSets = 1 + paoElementUniformBuffers.size();
+    tPoolInfo.maxSets = 1 + (uint)paoElementUniformBuffers.size();
 
     if (!ppLogicalDevice->CreateDescriptorPool(&tPoolInfo, nullptr, &poDescriptorPool))
     {
@@ -203,7 +203,7 @@ void cOverlayUniformHandler::CreateDescriptorSet(cScene* pScene)
     tAllocInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO;
 
     tAllocInfo.descriptorPool = poDescriptorPool;
-    tAllocInfo.descriptorSetCount = aoLayouts.size();
+    tAllocInfo.descriptorSetCount = (uint)aoLayouts.size();
     tAllocInfo.pSetLayouts = aoLayouts.data();
 
     if (!ppLogicalDevice->AllocateDescriptorSets(&tAllocInfo, &poDescriptorSet))
@@ -238,7 +238,7 @@ void cOverlayUniformHandler::CreateDescriptorSet(cScene* pScene)
     atDescriptorWrites[1].descriptorCount = 1;
     atDescriptorWrites[1].pBufferInfo = &tBufferInfo;
 
-    ppLogicalDevice->UpdateDescriptorSets(atDescriptorWrites.size(), atDescriptorWrites.data(),
+    ppLogicalDevice->UpdateDescriptorSets((uint)atDescriptorWrites.size(), atDescriptorWrites.data(),
                                           0, nullptr);
 
     std::vector<VkDescriptorSetLayout> aoElementLayouts(paoElementUniformBuffers.size(), poElementDescriptorSetLayout);
@@ -247,7 +247,7 @@ void cOverlayUniformHandler::CreateDescriptorSet(cScene* pScene)
     tElementAllocInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO;
 
     tElementAllocInfo.descriptorPool = poDescriptorPool;
-    tElementAllocInfo.descriptorSetCount = aoElementLayouts.size();
+    tElementAllocInfo.descriptorSetCount = (uint)aoElementLayouts.size();
     tElementAllocInfo.pSetLayouts = aoElementLayouts.data();
 
     paoElementDescriptorSets.resize(paoElementUniformBuffers.size());
@@ -287,7 +287,7 @@ void cOverlayUniformHandler::CreateDescriptorSet(cScene* pScene)
         atElementDescriptorWrites[1].descriptorCount = 1;
         atElementDescriptorWrites[1].pBufferInfo = &tElementBufferInfo;
 
-        ppLogicalDevice->UpdateDescriptorSets(atElementDescriptorWrites.size(), atElementDescriptorWrites.data(),
+        ppLogicalDevice->UpdateDescriptorSets((uint)atElementDescriptorWrites.size(), atElementDescriptorWrites.data(),
                                               0, nullptr);
 
         uiIndex++;
